Update patch set 2
Patch Set 2:

Can you please provide a reason on why 'git://' is the best default for everyone?

This patch is for fixing the case where 'git://' does not work.  To me this is an example of it not being a good default.  Can an example be provided that shows 'git://' working where 'https://' does not?

The problem with 'git://' is that with proxy servers it requires extra configuration to use a SOCKS proxy.  While 'https://' just works if the environment variable 'https_proxy=...' is set.

Patch-set: 2