Update patch set 14
Patch Set 14: Code-Review-1

I'd really like to see this done as an external plugin, especially one that has an uninstall phase. I think functionally this version of the code could be done that way easily in a stackforge project. If you spin one up I'll help with the conversion.

Patch-set: 14