Update patch set 2
Patch Set 2: Code-Review+2

I think this looks good.

The notion in one of the comments here needs to be re-enforced somewhere that since this is set to be called as a script and not sourced, any variables needed from stackrc/localrc need to be either exported or put on the command-line.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2