Update patch set 1
Patch Set 1: Code-Review-1

the fact that this isn't available from pypi is going to have breaking implications for all the world to isolate servers from clients in testing, because you won't be able to build a venv with this in it.

So I think the pypi piece needs to be figured out first, otherwise neutron puts itself in a position where it can't work like the rest of the system.

Patch-set: 1
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1