Update patch set 4
Patch Set 4: Code-Review-2

So as you can see from the many failures, the recursive nature of this causes problems ... especially when things like the pyOpenSSL package get upgraded over the top of the installed system package.

devstack has always had this uneasy balance between the system python packages defined in files/[apt|rpm] and what we install via pip.  it is not great, but it is what it is.  in a lot of cases this is required, as removing the system package can have flow-on effects of removing other required packages.  In other cases it's probably more true that it's "if it ain't broke..."

at a minimum i feel like this will introduce a lot more instability as we automatically drag in packages that have build dependencies that aren't met.

Perhaps you could give further examples of your environment and the issues you are hitting in the bug mentioned

Patch-set: 4
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2