Update patch set 3
Patch Set 3:

As you noted in the commit message, we've been back-and-forth on this a few times.  Configuration for proxies is not our default use-case.  What is it about setting GIT_BASE that does not work for you?

Patch-set: 3
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0