Update patch set 3
Patch Set 3:

Dean,

Because it doesn't work out of the box for people behind a proxy server.  And then when it fails a person has to spend most likely a lot of time trying to figure out about the configuration variable GIT_BASE.  So this is a known use case where the default breaks for people.

I work in a company with probably over 120,000 employees.  If any of them attempt to install Devstack out of the box, it fails for them.  Which likely gives them a poor user experience.

My confusion is that I have yet to hear any convincing argument on why 'git://' is superior to 'https:'.  If a case could be explained then fine, I would accept it.  Also if an individual person felt that 'git://' was preferable (which I find debatable), they could use the GIT_BASE variable.

On another note 'https://' appears to be less susceptible to a Man-In-The-Middle attack.

Thanks,
John

Patch-set: 3