Update patch set 31
Patch Set 31: Code-Review-1

(2 comments)

I'm still negative on this.  For a start it won't work with nounset because it assumes that things like RABBIT_* are defined, which they're not if the service is disabled.

I also don't really like how it is always building up this no_proxy and exporting it in environments where it isn't required.

the other thing is that this only seems to apply to things running under stack.sh ... if anyone is interacting don't they need to basically replicate this?

I think i could get on-board if

- it ran with nounset
- it didn't set no_proxy if it wasn't set
- it saved the no_proxy to a file that could be sourced
- included documentation about all this

Patch-set: 31
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1