Update patch set 5
Patch Set 5:

I'm hoping this can be a compromise option.  If https_proxy and HTTPS_PROXY environment variables are not set, then use 'git://' as the protocol as is currently done.

But if either 'https_proxy' or "HTTPS_PROXY' environment variables are set then use 'https://' as the protocol.

Patch-set: 5
Label: Verified=0