Update patch set 4
Patch Set 4:

Hi Ian, thanks very much for your explanation. I was very confused why such so 'obvious' change weren't implemented yet. So, IIUC from your explanation and the error on devstack, MySQL-python which is a new package can't find mysql_config, which I guess is not installed in the apt version of mysql-client/server, right?

Our use case, is that we want to speedup our CI by pre-installing all PIP dependencies on the VM image. But that fails after some time when a new project needs a newer version than the on in the image. Then, we need to constantly upgrade the image which will be complicated to automate.

Patch-set: 4