Update patch set 1
Patch Set 1:

OK, so you don't want a later version of this bringing in new dependencies on a stable branch, because those new/updated dependencies may break other projects?

it seems we have a couple of instances of this now, and they aren't consistent.  i think a get_from_requirements <package> function might be good

also style is to use $( over `

Patch-set: 1