Update patch set 1
Patch Set 1:

FWIW this was done as a hack during a previous wedge, stable/* branches got backports too.   Getting the version from requirements.txt is the Right Thing.

I have a feeling that we'll want to do this for the other top-level repos that may or may not be in project dependency lists (that's why OSC is installed this way, nothing else depends on it).

Patch-set: 1
Label: Verified=0