Update patch set 1
Patch Set 1:

This started as a wrapper around pip_install() and it made more sense to me to re-factor it in the end.  It is a separate function because we need to pass in the name/path of the venv and changing the signature of pip_install() did not seem like a good idea.  Maybe a -v option would have worked, we generally don't do that though except where those options are passed along to underlying commands.

pip_install() is (supposed to be) a low-level function, adding smarts to know which packages go into venvs seems to cross layers there.  I wanted to to be explicit at the install-this-code level what was going into a venv and where.  I'll rebase this and follow up with the next round to show how I was using it.

Patch-set: 1