Update patch set 2
Patch Set 2: Workflow-1

(1 comment)

This builds the client venv and installs a couple of client libs into it, not used by anything else yet.  There still seems to be some question about the process for installing the clients.

Also note that the clients are hard-coded to the venv, this is not optional at this point, does it need to be?

Patch-set: 2
Label: Workflow=-1