Update patch set 3
Patch Set 3:

At this point I want to be specific on which command is being called.  The follow-on to this (not proposed yet) fully-qualifies the client calls.  In the end that may not be necessary, but during the transition to full service venvs I think it is to ensure that the correct clients are being used.

I clearly need to debug this a bit more before switching things...

Patch-set: 3
Label: Verified=0