Glance - configure keystone_authtoken separately
The move to configuring auth_token middleware by auth plugin
(in https://review.openstack.org/#/c/142967) is not yet supported by
glance (in the same way heat was changed in that patch).

This patch sets keystone_authtoken for glance-registry and glance-api
as it was configured prior to that patch. Glance should be updated to
support the new configuration settings.

Change-Id: Ic53fc468305f9fa1d44cae50e7dcebf6cd2eed39
Closes-Bug: 1421467