Update patch set 7
Patch Set 7:

At this point this is working for our non-Grenade jobs.  We will turn off USE_CLIENT_VENV to merge this so it will be benign until enabled in devstack-gate.  Once it merges Grenade will be able to set up the client-venv and the tests pass.

Patch-set: 7
Label: Verified=0