Change config of keystone_authtoken in lib/ironic
Ironic re-uses specific values from the keystone_authtoken middleware.
So we should specify admin_user, admin_password and so on.
Until the function(bugs#1418341) of status of 'Won't fix' is implemented,
these statement must be remained.

The other bug affecting devstack in bugs#1418341 is resolved by this commit and closed it.
reffer to https://bugs.launchpad.net/ironic/+bug/1418341

Change-Id: If8869e16a167eea0af87afda0eabcbb803627db6
Closes-Bug: #1418341 at devstack