Update patch set 1
Patch Set 1: Code-Review-1

This really belongs in tools/fixup_stuff.sh as it is correcting/working around a Python packaging problem and not part of installing pip.  There is a section there for Python Packages.

Also can you verify if this is a problem for all supported distros?  I'd hate to fail on a distro different from yours because a package name is different for example.

Patch-set: 1
Label: Code-Review=-1