Update patch set 3
Patch Set 3: Code-Review-1

This looks to be reverting ironic to using the old style of configuring auth_token middleware. The bug references that the options do not match the documentation, unfortunately that is a doc bug in middleware. 

I am trying to determine what is going on here, but this looks like the wrong approach.

Patch-set: 3
Reviewer: Gerrit User 2903 <2903@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1