Update patch set 3
Patch Set 3:

(Sorry for the multiple comments). 

The fix here looks to solve the issue for ironic, but does so by reverting the auth_token config instead of updating how ironic gets the data to interact with keystone.

This might be needed as a stop-gap until ironic is updated/fixed.

Patch-set: 3