WIP: Enable setting "[volume] volume_size" in tempest.conf
Devstack wasn't able to set the tempest variable "volume_size" in the
[volume] section of the tempest configuration. The default size of 1GB
couldn't be changed then. This could lead to an issue when the guest
image size is bigger than that.
This patch fixes that by providing a new variable which sets the
volume_size in the tempest.conf file.

Closes-Bug: 1422396
Change-Id: I8053553a216636a5bd06d42a7aa2ebfff4a963e5