Definition of _default _member_ role is necessary
The role _member_ was previously implicitly created passing the
parameter --project when a user was created in Keystone. This behaviour
was changed in the commit https://review.openstack.org/#/c/150667/ so
the role _member_ is not anymore created in Devstack, but without its
existence it is impossible to create or edit projects.
Rather than reverting the commit that caused this bug, this patch
explicitly create the role _memeber_ even if it is not assigned to any
user.

Change-Id: I5f822b827e25a645bbb41986cef35939812a6908
Closes-Bug: 1421616