Definition of _default _member_ role is necessary
The role _member_ was previously implicitly created passing the
parameter --project when a user was created in Keystone. This behaviour
was changed in the commit https://review.openstack.org/#/c/150667/ so
the role _member_ is not anymore created in Devstack, but without its
existence it is impossible to create or edit projects in Horizon.
This patch restore --project as a parameter of get_or_create_user,
optional this time, and pass it to the function only when the user demo
is created, so the default _member_ role will be added in keystone.

Change-Id: I5f822b827e25a645bbb41986cef35939812a6908
Closes-Bug: 1421616