Update patch set 3
Patch Set 3:

The Blog post that helps outline how to do all of this has disappeared. The documentation here: has most of what you would need, but the docs are not very friendly.

This might be easier to fix in devstack with a big TODO and reopen the ironic bug.

Patch-set: 3