Don't create stack-volume-default in lib/cinder
In the past lib/cinder would always call
init_default_lvm_volume_group which would use the
/lib/lvm default VG name of stack-volumes-default.
We switched over to using the be_type and be_name
to create and use stack-volume-lvmdriver-X, but we
now have both calls in place.

The result is that we have two VG's, one on the loop
device named stack-volumes-default which actually
isn't used by the Cinder tests, and another on
/dev/vdX that's actually the intended Cinder VG.

To reiterate, the current gate tests are in fact only
using the stack-volumes-lvmdrver-X VG.

Closes-Bug: #1422532

This may be in order to preserve compat for Grenade tests, but we should probably do something a bit better to keep from having an unused VG in the standard dsvm full run to eliminate some confusion.

Change-Id: I8e8b40f1d0d1ad7be0f9abf210ab1ea52c7468a3