Update patch set 3
Patch Set 3:

@Dean Troyer:
> How would this be used that can not be handled by setting 
> directly from local.conf when required?
In a CI environment devstack is used to create a VM which gets tested by tempest. The configuration for tempest is automatically generated by devstack too. Manually changing the automatically generated "tempest.conf" every time the CI environment has to check an incoming change seems to be undesirable. With this patch set it is possible to create a "local.conf" for devstack once, which is used for multiple CI executions.

Patch-set: 3