purge all pure python libraries
python libraries should be installed from upstream, not from the
distro, as much as possible. The following is a first attempt at
purging all the python libraries and making it so that they instead
should fall back to the pypi versions.

Libraries which are known to include native code are left behind.

Change-Id: I47b7e787771683c2fc4404e586f11c1a19aac15c