Update patch set 1
Patch Set 1:

Thank you for the comments, I think I understood the problem. When in test keystone is required to add a user to a tenant, it looks for the default _member_ role with the default member_role_id "9fe2ff9ee4384b1894a90878d3e92bab", but cannot find it so try to create the _member_ role, but it fails cause it exists yet.

This is due to the fact that specific id, defined in keystone config file,  is assigned to the _member_ role when it is implicitly create it adding a user to a project within a tenant.

In my case the _member_ role is explicitly created with get_or_create_role method, so it will assume a role_id different from the default one, causing the fails in the tests.

Patch-set: 1