Update patch set 1
Patch Set 1:

openstack command line doesn't give the chance to specify an role_id when creating a new role.

I think the easiest way to create the _member_ role is to give again to get_or_create_user a parameter "project", but optional this time and to pass it only when creating the user "demo", so we will have a demo user with role _member_ (and Member too) of tenant demo. Alternatively we can do the same when creating the admin user instead of demo user.

It is to be noted that the only thing that matter here is the existence of the _member_ role, not the fact it is assigned to a specific user, anyway the only way (or anyway the easiest way) to create the _member_ role is to pass the project (tenant) parameter when creating a user.

I still don't understand WHY Horizon needs a _member_ role, anyway I don't feel confident to change that behaviour without a more deep discussion.

Patch-set: 1