Update patch set 7
Patch Set 7: Code-Review-1

(1 comment)

So I think at a min. the wording needs to be updated to make it a bit more clear.

I'm still on the fence about this, I disagree with clarkb's assertion in the d-g review that we should make the ci optimized case the default in devstack. Doing setup_develop on tempest is the sane default for tempest dev. But, I don't feel strongly enough about it either way. So if we update the wording I think this'll be fine.

Patch-set: 7
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1