Update patch set 2
Patch Set 2: Code-Review+1

Looks good. But this would make things contrary to the patch that was merged earlier (user gains token scoping information
for a project). May be should we create and delete a user for this purpose. May be it is not a good way but just thinking of an alternative.

Patch-set: 2
Label: Code-Review=+1