Update patch set 2
Patch Set 2:

@Ian
Suppose that's fine but quite frankly I don't know why that should be a part of this patch.  This patch is explicitly to remove the call to the create in cinder setup.  Frankly I don't care if there's an unused method left behind, but think it's fine to pull that out as a separate patch.

If this is really important for some reason I can amend this change to remove it entirely but personally that doesn't seem like the right thing.

Patch-set: 2