Update patch set 2
Patch Set 2:

@Ian
So turns out Dean had some insight that the "shared nova/cinder" add was fairly recent to try and reduce the disk usage on test instances.  Problem is that the way the Cinder lib is right now it actually has the opposite effect and creates a VG that is never used.

There is work that needs to be done here to clean all of this up and sort it out a bit better, however from the perspective of this change it really is unrelated IMO.  I'm perfectly fine if this doesn't merge, it doesn't hurt anything as it is now but it does cause some confusion and is creating a VG that is never used (at least by the gate).

Patch-set: 2