Install python-redis from pip in Ubuntu
Zaqar depends on redis' sentinel. This feature is shipped in the
Ubuntu's redis-server package. However, the python-redis package is
older than what Zaqar needs, therefore we must install it from pip for
Zaqar to work correctly.

This is a temporary fix until the package is updated and/or Zaqar's
implementation is re-written.

Change-Id: Idfef2fe1b41b44c4cba6b6948b621bc44e77659b