Support for single interface Neutron networking
When running Neutron on a single node, that only has a single interface,
it is nessesary to take the IP address from the physical interface, add
the interface to the OVS bridge, add the IP address to the OVS bridge,
then update the routing table.  The reverse is done on cleanup.

Change-Id: Ie35cb537bb670c4773598b8db29877fb8a12ff50