Do Not Merge: Test test-accounts with roles
This commit generates a test accounts file with roles and configures
tempest to use the locking account creds provider instead of tenant
isolation. Things are hard coded so this isn't really a candidate for
merge in this state. This creates an accounts.yaml with 8 regular
user creds, 8 admin users, 4 users with the swift operator role, and
4 users with the swift reseller_admin role.


Depends-On: I07fa785ef0bad9677885488f3155df7d47f8c846
Change-Id: I5fe09fdae2fb8082cb97fbcc46302939b328afbb