Do Not Merge: Test test-accounts with roles
This commit generates a test accounts file with roles and configures
tempest to use the locking account creds provider instead of tenant
isolation. Things are hard coded so this isn't really a candidate for
merge in this state. This creates an accounts.yaml with 8 regular
user creds, 8 admin users, 4 users with the swift operator role, and
4 users with the swift reseller_admin role.

This omits setting a network devstack change:
I5fe09fdae2fb8082cb97fbcc46302939b328afbb tests using neutron to set
a network. It was easier to push a second patch that fix the if logic
in that change.

Change-Id: I76006bda934320e5ed7d2a694be50164ad8edc33
Depends-On: I82af7a3cba1e62cada94a9f4220666f1733e0575