Adding quotation marks in order to parse string correclty on RHEL7
Problem:
On rhel8 the lvm.conf filter is generated wrongly

Root Cause:
RHEL7 seems to behave a litle differnt regards the arguments after the
"<<<" operator. To interpret the whole variable as a single string,
quotation marks around the variable are required. Ubuntu can deal
with both ways.

Resolution:
This commit adds the required quotation marks to the arguments after
the "<<<" operator

Closes-Bug #1425839

Change-Id: Ie82fcd7ef0deacbb6aaf18c5c466a5d5baf52681