Update patch set 2
Patch Set 2: -Workflow

Yeah it looks odd, I'm trying to figure out how to simplify but the thing is that it's the situation is complicated.

is_deployed_by_agent looks if the node has the agent_* driver, which requires things like Swift Temp URLs to be able to deploy it. Now the same ramdisk that is used for the agent_* drivers can also be used by the other drivers (it's pluggable) but without all the swift bits.

We are doing a forklift in Ironic, we don't want to keep supporting 2 different deploy ramdisks, so we are moving the drivers to use the ramdisk that the agent uses, because it's more flexible and pluggable. But we still have to keep compatibility with the current default one for a while, so people have time to migrate.

So in the future, when we have the agent ramdisk (called IPA) as the default and only one supported, we can remove all those conditions where "is_deployed_by_agent || IRONIC_DEPLOY_ISCSI_USE_IPA" is being checked. And leave the ones where only "is_deployed_by_agent" is tested.

Does it makes things a bit more clear about the situation? Suggestions are welcome too.

Thanks

Patch-set: 2