Update patch set 3
Patch Set 3:

@afazekas, plz have a look at my comment on the 4th of feb :

"I still think my patch is valid because, when your patch to oslo-incubator gets in, it will take months before every project sync up with the new version of oslo-incubator. In the mean time, the logs have hundreds or thousands of not that useful messages."

It's already fixed upstream. But the still affects/and will affect stable branches.

Patch-set: 3