Adds networking_hyperv as a neutron third party driver
This commit adds the configuration file necessary for
devstack to be able to clone the stackforge repo for
the project. In order to do so, users will have to add
the following line in their local.conf file:

enable_service networking_hyperv

This commit is useful for users that have configured
devstack to use the ML2 neutron plugin and 'hyperv' as a
mechanism driver and wish to be up to date with the changes.
This is particularly useful for the Hyper-V CI.

Stackforge repo: https://github.com/stackforge/networking-hyperv

Change-Id: Ice93371c39abac5e3783f988ac38865bcb471294