Create _member_ role in keystone bootstrap
The _member_ role must be created because horizon uses this as the default role
to add to users in projects. Unfortunately because we need to add the role id
to the CONF file we have to create this role, insert it into the config file
and then restart keystone to have the change take effect.

This is a similar process as what needs to happen with defining administration
domain_ids for policy so create a function that can be used to bootstrap
keystone with these sort of database/config file synchronized values.

Change-Id: I65738b4f52e750abfc34995e2d618f9b1c57e5b3
Closes-Bug: #1421616