Update patch set 1
Patch Set 1: Code-Review-1

I don't agree with your approach, since there is a much easier way to create a _member_ role that is accepted by Keystone.

When you use openstack CLI to create an user and pass the parameter --project, than to the newly created user is assigned a _member_ role for that project and - what matters for us - a _member_ role is created if it doesn't already exist.

This was the procedure followed before this merge https://review.openstack.org/#/c/150667/ and this is what I proposed to have again in this bug fix proposal https://review.openstack.org/#/c/156527/

Patch-set: 1