Update patch set 9
Patch Set 9: Code-Review-1

This looks like a good step in the right direction. I'm curious if you have plans to also add the reverse of this. On a "unstack.sh", do you plan to remove the IP from the bridge and put it back on the single interface?

Patch-set: 9
Label: Code-Review=-1