Update patch set 2
Patch Set 2:

Pasquale: I don't necessarily disagree with you. This is a crappy fix for something that is really a keystone bug. I assume most deployments must just do enough with default_project_id that it doesn't come up - because i didn't know about it.

My main problem with letting keystone build it is that (as discovered) it is really non-obvious why everything fails when the role is not present. If we go down that route i'd like to see a bit comment section on why we _have_ to have at least one user with a default_project_id.

Horizon is configured to require _member_ as the default value for when you add users to a group. I'm guessing what happens is that horizon does a lookup of the _member_ role name, gets the id and then uses the id from then on. When I added the _member_ role manually (id still set wrong) horizon was working - just other CI systems fail apparently.

Patch-set: 2
Label: Verified=0