Add a simple memory/process tracking service
We are experiencing OOM issues with centos/cinder jobs which are
turning out to be quite difficult to replicate anywhere but in the
gate on rax hosts.

Although we currently have 'dstat' running through the run, this
doesn't always give you the whole view.  It doesn't show important
things like what vm's have started up, or what processes might be
hanging around stuck in various states.

Logging this data isn't appropriate for every job, but it would be
really nice to be able to get much better insight into such an issue
like this by simply turning on a service.  It would avoid nodepool
root-users needing to be around to do this sort of poking on
everyone's behalf by logging into hosts.

This adds a memtracker service which collects /proc data, ps data
including memory usage, number of threads, wait-channel, cpu-time and
command-line in a tight loop.

Also added is a very simple analysis script

Change-Id: Id622a818e3d98eabf014971f46ed66a6a63b64bb