Update patch set 4
Patch Set 4:

afazekas: The problem is that _member_ is a legacy of the v2->v3 conversion process. It used to be that a user was a member of a project, now membership is defined by having a role in that project. 

I don't expect keystone will db_sync create a role that is not used in v3. There are no other roles that are created by default as defining them has never been considered keystone's jurisdiction.

Also, as mentioned, i agree there is a keystone bug here. We need to work around it until it is fixed. Also as mentioned in the review the same issue is present in upcoming policy changes so the restart is going to have to happen one way or another.

Patch-set: 4
Label: Verified=0