Remove spaces from extension lists
This commit removes all of the spaces from the extension lists in the
env vars for passed into iniset. With spaces these lists are only getting
the first item set in the output tempest.conf file which means we only run
a very small subset of the tests we intend to.

Change-Id: I489f6ec4dd5b1824ebff576902746b160c6b8bc5