Update patch set 1
Patch Set 1: Code-Review+2

LGTM, we might have allowed in a tempest test which doesn't work on the sideways job. If so we'll have to land a tempest change to exclude it in that scenario and mark this as depends-on for that patch.

Patch-set: 1
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2