Update patch set 3
Patch Set 3:

ayoung I would agree with you, but Horizon dashboard requires the existence of the role _member_ otherwise it gives an error when try to manage projects.

We all agreed this is not really a devstack bug and both mine and the other one proposed are just workaround to make things work while the real bug is corrected.

Anyway I am still not sure the bug is in keystone as other proposed and not actually in horizon... why should horizon require a _member_ role?

Patch-set: 3