Set DevStack to dual stack by default in Kilo+
This patch sets DevStack to run in dual stack networking, with both IPv4
and IPv6 networking configured. This change is required for dual stack
testing at the gate.

A different patch[1] was created against devstack-gate to make this the
default, but the Juno branch of Neutron is missing required fixes to the
L3 agent that are present in Kilo. This was the suggested alternative.

Related-change: I3d416275f77913769b98e77f7e47bed17fc4d1cc

Change-Id: If0e0b818355e4cb1338f7fa72af5e81e24361574
Depends-On: I150da5938e79eeef0c947ddb1a4282e37d0515ee