Update patch set 12
Patch Set 12:

The commit message starts with 'When running Neutron on a single node'.

I still does not get the exact use case. If you do not have permissions on the router for adding a route entry, why you don't do it on the clients ? (For example on a tempest node?)

If you do not have access to the router/switch you had to have an allocated ip-range `per single` node, it is also something not easy to manage.

Can you be more specific in the commit message what exactly you want to achieve ?

Patch-set: 12