Update patch set 12
Patch Set 12:

Gary, the reasoning behind having this document here is so that users will be able to configure neutron-lbaas without knowing where to find it.  Even though the devstack plugin has been moved, it seems to make sense to have a central place to look for installation documentation.  Al

Patch-set: 12
Label: Verified=0