Set my_ip in cinder.conf with value of HOST_IP
The current issue is that if we deploy c-vol service on a separate
machine, my_ip and SERVICE_HOST will be different, because my_ip is
the machine where c-vol service is running and SERVICE_HOST points
to the machine where the cinder api service is running. If my_ip of
c-vol in cinder.conf is set to the IP of c-api, it will cause the
issue that the volume is unable to attach. So we need to set the my_ip
to HOST_IP intead of SERVIVE_HOST.

Change-Id: I699c0b5297c60e9f9934f74684abf563f4b0e977
closes-bug: #1428013