Do not uninstall the python-virtualenv
The files/rpms/general:python-virtualenv explicitly
installed this package on Fedoras.

If you stack/unstack the package gets reinstalled,
and confuses the system about the installed virtual-env version.
The uninstall works in CI, but it can be problematic when you do
reinstalls on the same machine.

The uninstall introduced by
834b804d3eda9029d3c66db0ab732a76a22ed08b, this commit deos not
has any reference to the external bug what it supposed to solve.

I think both of us knows the rpm vs. pip issues.
Usually `safer` to handle this issues by installing the
rpm and overriding it with pip than not installing the rpm.

Related RDO thread started here:
https://www.redhat.com/archives/rdo-list/2015-March/msg00015.html

Change-Id: I4a723f179bdc28d39a4910fb9e3787e9e67c354b