Update patch set 1
Patch Set 1: Code-Review-1

It would appear then that the real fix is to remove python-virtualenv from files/rpms/general and not install it in the first place.  Actually that was my oversight in the original change.  At the _very_ least, this is a problem only for the is_fedora() distros and should be restricted to them.

The situation with pip-installed packages on RH distros is broken.  I don't have the history of why it has never been fixed, I don't care.  We've had to work around that brokenness more than once and it is a pain to debug.

Having a package installer fight dropping files on top of each other is not a solution.  Removing one of those package managers from the picture is the only solution that works in the current RH environment.

site-packages vs dist-packages.  Jeez, even Perl4 got this right on Windows back in 1995...

Patch-set: 1
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1