Update patch set 2
Patch Set 2: Code-Review+2

1) the original commit should have probably been gated with some check.  it's not going to be true for perpetuity that the bug will exist in the python-virtualenv package

2) it doesn't really matter, i don't see any need to install the packaged version, so this looks good

3) yes we've discussed this pip/rpm issue before.

a) i believe things are changing in fedora so that packages are built with pip, so essentially pip can overwrite them correctly.  the issue can be at the moment that pip doesn't know how to remove everything related to a package, so if an old file is left around sometimes it plays havoc.

that's why we get

 "Uninstalling a distutils installed project (...) has been deprecated and will be removed in a future version."

i believe in the future pip will know all the files installed by the rpm, get rid of them and start again.

b) not that this really helps centos7; but we really are targeting a completely blank-install, so there's a limited number of python packages that *have* to be installed (things relating to cloud-init).  we really could/should strip out a lot of the package installs i think and reduce our exposure to the problem.

Patch-set: 2
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2