Update patch set 12
Patch Set 12:

"But I have to admit, I am running Ubuntu ;)"

touche :)

I can happily report that I have that devilish NetworkManager crap turned off.

When you say "devstack runs in a VM" are you saying "I can get it to work both on the laptop and on the vm" or "I only do devstack in the VM and it works fine".

If its the latter, that's the difference. The goal here is baremetal devstack with just one physical nic.

If that's what you're doing then awesome for you but I wonder what the difference is.

Patch-set: 12