Update patch set 12
Patch Set 12:

Chris, I meant I can run devstack on my laptop ("baremetal" with one physical network interface), it works ( both outgoing connection to the Internet and incoming connection through a floating IP from my laptop). And it also works when I do the same setting from a VM hosted on a public cloud.

But I am far from a network guru, and I know this stuff can be complicated so I am not going to have an opinion on this patch :)

Patch-set: 12