add the port_sec as default neutron/ml2 extension driver
Neutron ML2 plugin introduces the first extension driver port_security, this
patch add it to be a default extension driver as a example. And also, if not
set it by default, networks like public/private which are created after the
neutron-db-manage's update, will not include the port-sec value.

Note: If we comment the extension_drivers in ml2_conf.ini, port_security will
be loaded. If we set the extension_drivers to NULL, none extension driver will
be loaded.

Change-Id: I3035317c83d22804855517434bd8578719ce0436
Depends-On: I2da53168e2529db7a8094ce90ef3a8a93fe55727