Update patch set 2
Patch Set 2:

pgsql was changed for specific reasons that (at the time) did not apply to mysql.  "Because $OTHER_DB has it" is not a reason, what is the specific reason?

The fact that they might be different is OK in this case, it isn't an exact 1-to-1 comparison.

Patch-set: 2